datasources/requirements.txt
kafka-python==2.0.2
avro-python3==1.9.2.1
json-numpy==2.1.0